Add support for Response trailers for netstandard20
When using the netstandard20 version, response trailers were
previously ignored. This was causing issues with handling
responses from a GRPC service.

Grpc-dotnet works with WinHttpHandler on recent Windows versions, and
trailers are stored in the request generic property bag as shown here:
https://github.com/grpc/grpc-dotnet/blob/456098bcfeb9c796c19ea023b3ab3703181f1089/src/Shared/TrailingHeadersHelpers.cs#L30

This PR uses the same way to store trailers when compiled for netstandard 2.0

Signed-off-by: Simon Ferquel <simon.ferquel@unity3d.com>